+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/AlbumEditPanel.java
@@ -27,7 +27,6 @@ import is.codion.swing.framework.ui.EntityEditPanel;
 
 import javax.swing.JPanel;
 import java.awt.BorderLayout;
-import java.util.Collections;
 import java.util.List;
 
 import static is.codion.demos.chinook.domain.api.Chinook.Album;
@@ -77,7 +76,7 @@ public final class AlbumEditPanel extends EntityEditPanel {
 		// We create JList based value for the album tags.
 		ComponentValue<List<String>, FilterList<String>> tagsValue =
 						createList(FilterListModel.builder()
-										.items(Collections.<String>emptyList())
+										.<String>items()
 										.build())
 										// The value should be based on the items in
 										// the list as opposed to the selected items